Add test for LockFileEx failure (#4)
This PR adds a test for LockFileEx failure.

I saw an issue in using this package that felt like there was a bug in
relaying LockFileEx failure. Prior to this PR, there were no tests for
LockFileEx failure, so I added a test.

This test passes, so the bug I observed must be elsewhere. But let's
keep this new test.